+++ src/main/java/org/flag4j/arrays/backend/ring_arrays/AbstractDenseRingVector.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2024. Jacob Watters
+ * Copyright (c) 2024-2025. Jacob Watters
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -90,7 +90,7 @@ public abstract class AbstractDenseRingVector<T extends AbstractDenseRingVector<
      *
      * @return The Euclidean norm of this vector.
      */
-    public double norm(int p) {
+    public double norm(double p) {
         return VectorNorms.norm(data, p);
     }
